Extract queue ad runtime into useQueueAdRuntime hook
